fix: Use initialState prop instead of useParams in installation-page
The installation page is composed of multiple steps (presentation,
security, hint, configureExtension)

When accessing the page, we want to be able to specify in the URL the
step that should be displayed

This is used in a specific scenario where we redirect to
cozy-settings's change password page. When the password is changed,
then cozy-settings should redirect to the step that was displayed
before

In previous implementation, we used `useParams` from `react-router`
(legacy from old cozy-password only website). But as we are displaying
this component from an Angular website, then the routing is not handled
from React anymore

Instead of trying to make `useParams` work, we chose to inject an
`initialStep` prop from the Angular context. Then the URL param reading
would be done from the Angular context

Also since react-router is not used anymore, then the InstallationPage
step is not synchronized anymore with the URL. So instead of using a
path (i.e. `installation/hint`), we chose to convert this param to a
QueryParams (i.e. `installation?initialStep=hint`), this would prevent
any confusion