fix: Navigate to home and other apps from cozy-bar in flagship app
The new cozy-bar added recently leverages the WebviewIntentProvider to navigate to home and other apps on flagship app.

Unfortunately, it is not possible to use multiple WebviewIntentProvider in an app. cozy-intent does not support it. But the cozy-pass-web app is built around multiple React wrapper so we can not have one WebviewIntentProvider for the entire app and we also need it to check if the pass mobile app is installed during the onboarding process.

So here we :
- move the WebviewIntentProvider from ReactWrapper to the BarWrapper so it is instanciated only once
- expose a proxy using JS events to allow other part of the code to call webviewIntent.call
- use this proxy to check if the pass mobile app is installed